requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
